+++ CHANGELOG.md
@@ -8,4 +8,4 @@
 **Improvements**
 - Added Craft Requirement tooltip to Recipes
   - If a recipe has a craft requirement, it will now display a small exclamation and tooltip when you hover over it in RRV.
-- 
+- Added an option to hide Category Buttons.